Refactor based on requests on Review
